some refactoring
- move LocalNode#lookup() to Sim class
- NetworkParams.NETWORK_TIMEOUT -> ACK_TIMEOUT
- add Network.Params.REPLY_TIMEOUT

+++ piax-gtrans/src/main/java/org/piax/gtrans/async/EventExecutor.java
@@ -310,7 +310,7 @@ public class EventExecutor {
         latencyProvider = p;
     }
 
-    public static int latency(Node a, Node b) {
+    public static long latency(Node a, Node b) {
         if (EventExecutor.realtime.value()) {
             return 0;
         }
